moving things to kotlin, fixing toast for Kotlin, moving things here and there for MVVM
