+++ app/src/main/res/layout/custom_toast.xml
@@ -11,7 +11,6 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_gravity="center_horizontal"
-
         android:padding="12dp"
         android:textColor="@android:color/white"
         android:textSize="14sp" />
