+++ app/src/main/res/layout/fragment_second.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.searoth.template.fragment.SecondFragment">
+    tools:context="com.searoth.template.viewmodel.main.fragment.SecondFragment">
 
     <android.support.constraint.ConstraintLayout
         xmlns:android="http://schemas.android.com/apk/res/android"
@@ -10,7 +10,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context="com.searoth.template.fragment.SecondFragment">
+        tools:context="com.searoth.template.viewmodel.main.fragment.SecondFragment">
 
         <TextView
             android:id="@+id/textView4"
