+++ app/src/main/res/layout/fragment_third.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.searoth.template.fragment.ThirdFragment">
+    tools:context="com.searoth.template.viewmodel.main.fragment.ThirdFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
